+++ tests/testthat/test-linreg.R
@@ -229,3 +229,22 @@ test_that("calculation of residuals for 3d arrays works", {
     expect_equal(expected, resid)
 
 })
+
+test_that("linreg full variance-covariance matrix is correct", {
+
+    set.seed(20260608)
+    n <- 100
+    x <- cbind(1, rnorm(n))
+    y <- x %*% c(1,1) + rnorm(n)
+
+    out_lm <- summary(lm(y ~ -1 + x))
+
+    out_var <- fit_linreg_eigenchol(x, y, FALSE, TRUE)
+
+    out_se <- fit_linreg_eigenchol(x, y, FALSE, TRUE)
+
+    expect_equivalent(out_se$SE, out_var$SE)
+    expect_equivalent(out_var$SE, out_lm$coefficients[,2])
+    expect_equivalent(out_var$var, out_lm$cov.unscaled * out_lm$sigma^2)
+
+})